Model/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
aniso8601=9.0.1=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
audioread=2.1.9=pypi_0
blas=2.113=mkl
blas-devel=3.9.0=13_win64_mkl
brotlipy=0.7.0=py38h294d835_1003
bzip2=1.0.8=h8ffe710_4
ca-certificates=2021.10.8=h5b45459_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py38haa244fe_2
cffi=1.15.0=py38hd8c33c5_0
charset-normalizer=2.0.12=pyhd8ed1ab_0
cheroot=8.6.0=pyhd8ed1ab_0
click=8.1.1=py38haa244fe_1
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py38h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.0.0=pypi_0
decorator=5.1.1=pypi_0
dill=0.3.4=pypi_0
faiss=1.7.2=py38cuda112h79539ec_0_cuda
faiss-gpu=1.7.2=h949689a_0
filelock=3.6.0=pyhd8ed1ab_0
flask=2.1.1=pypi_0
flask-cors=3.0.10=pypi_0
flask-restx=0.5.1=pypi_0
flask-wtf=1.0.1=pypi_0
fonttools=4.31.2=pypi_0
frozenlist=1.3.0=pypi_0
fsspec=2022.3.0=pypi_0
gevent=21.8.0=pypi_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=1.1.2=pypi_0
grpcio=1.45.0=pypi_0
huggingface-hub=0.4.0=pypi_0
huggingface_hub=0.0.17=py_0
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.6.0=pypi_0
inflect=5.4.0=pypi_0
intel-openmp=2022.0.0=h57928b3_3663
itsdangerous=2.1.2=pypi_0
jaraco.functools=3.5.0=pyhd8ed1ab_0
jieba=0.42.1=pypi_0
jinja2=3.1.1=pypi_0
joblib=1.1.0=pyhd8ed1ab_0
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jsonschema=4.4.0=pypi_0
kiwisolver=1.4.2=pypi_0
libblas=3.9.0=13_win64_mkl
libcblas=3.9.0=13_win64_mkl
libfaiss=1.7.2=cuda112h0f75059_0_cuda
libfaiss-avx2=1.7.2=cuda112h1234567_0_cuda
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=13_win64_mkl
liblapacke=3.9.0=13_win64_mkl
librosa=0.9.1=pypi_0
libuv=1.43.0=h8ffe710_0
libzlib=1.2.11=h8ffe710_1014
llvmlite=0.38.0=pypi_0
loguru=0.6.0=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
mkl=2022.0.0=h0e2418a_796
mkl-devel=2022.0.0=h57928b3_797
mkl-include=2022.0.0=h0e2418a_796
more-itertools=8.12.0=pyhd8ed1ab_0
msys2-conda-epoch=20160418=1
multidict=6.0.2=pypi_0
multiprocess=0.70.12.2=pypi_0
nltk=3.7=pypi_0
numba=0.55.1=pypi_0
numpy=1.22.3=py38hcf66579_0
oauthlib=3.2.0=pypi_0
openssl=3.0.2=h8ffe710_1
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=pypi_0
patsy=0.5.2=pypi_0
pillow=9.0.1=pypi_0
pip=22.0.4=pyhd8ed1ab_0
pooch=1.6.0=pypi_0
protobuf=3.19.4=pypi_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pynndescent=0.5.6=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pypinyin=0.46.0=pypi_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.9.1=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38haa244fe_4
python=3.8.13=hcf16a7b_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pyyaml=6.0=py38h294d835_3
pyzmq=22.3.0=pypi_0
regex=2022.3.15=py38h294d835_0
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
resampy=0.2.2=pypi_0
responses=0.18.0=pypi_0
rsa=4.8=pypi_0
sacremoses=master=py_0
scikit-learn=1.0.2=pypi_0
sentence-transformers=2.2.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=61.2.0=py38haa244fe_3
six=1.16.0=pyh6c4a22f_0
sounddevice=0.4.4=pypi_0
soundfile=0.10.3.post1=pypi_0
sqlite=3.37.1=h8ffe710_0
statsmodels=0.11.0=pypi_0
tbb=2021.5.0=h2d74725_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
text2vec=1.1.5=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h8ffe710_0
tokenizers=0.10.3=py38_0
torchfile=0.1.0=pypi_0
torchvision=0.12.0=pypi_0
tornado=6.1=pypi_0
tqdm=4.63.1=pyhd8ed1ab_0
transformers=4.12.5=pyhd8ed1ab_0
typing-extensions=4.1.1=hd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
ucrt=10.0.20348.0=h57928b3_0
umap-learn=0.5.2=pypi_0
unidecode=1.3.4=pypi_0
urllib3=1.26.9=pyhd8ed1ab_0
vc=14.2=hb210afc_6
visdom=0.1.8.9=pypi_0
vs2015_runtime=14.29.30037=h902a5da_6
web.py=0.62=pyhd3deb0d_0
webrtcvad=2.0.10=py38hbd9d945_1
websocket-client=1.3.2=pypi_0
werkzeug=2.1.0=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
win32-setctime=1.1.0=pypi_0
win_inet_pton=1.1.0=py38haa244fe_3
wtforms=3.0.1=pypi_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=h8ffe710_2
yarl=1.7.2=pypi_0
zipp=3.7.0=pypi_0
zope-event=4.5.0=pypi_0
zope-interface=5.4.0=pypi_0
